+++ src/Centreon/Domain/AccessControlList/Interfaces/AccessControlListRepositoryInterface.php
@@ -23,7 +23,7 @@ declare(strict_types=1);
 namespace Centreon\Domain\AccessControlList\Interfaces;
 
 use Centreon\Domain\Contact\Interfaces\ContactInterface;
-use Centreon\Domain\Security\AccessGroup;
+use Core\Security\Domain\AccessGroup\Model\AccessGroup;
 
 interface AccessControlListRepositoryInterface
 {
