+++ src/Centreon/Domain/Monitoring/HostGroup/HostGroupService.php
@@ -25,7 +25,7 @@ namespace Centreon\Domain\Monitoring\HostGroup;
 use Centreon\Domain\Service\AbstractCentreonService;
 use Centreon\Domain\Monitoring\HostGroup\Interfaces\HostGroupServiceInterface;
 use Centreon\Domain\Monitoring\HostGroup\Interfaces\HostGroupRepositoryInterface;
-use Centreon\Domain\Security\Interfaces\AccessGroupRepositoryInterface;
+use Core\Security\Application\Repository\AccessGroupRepositoryInterface;
 
 final class HostGroupService extends AbstractCentreonService implements HostGroupServiceInterface
 {
