+++ src/Centreon/Domain/Monitoring/Timeline/TimelineService.php
@@ -24,7 +24,7 @@ namespace Centreon\Domain\Monitoring\Timeline;
 
 use Centreon\Domain\Monitoring\Timeline\Interfaces\TimelineServiceInterface;
 use Centreon\Domain\Monitoring\Timeline\Interfaces\TimelineRepositoryInterface;
-use Centreon\Domain\Security\Interfaces\AccessGroupRepositoryInterface;
+use Core\Security\Application\Repository\AccessGroupRepositoryInterface;
 use Centreon\Domain\Service\AbstractCentreonService;
 use Centreon\Domain\Monitoring\Host;
 use Centreon\Domain\Monitoring\Service;
