+++ src/Centreon/Infrastructure/Monitoring/ServiceGroup/ServiceGroupRepositoryRDB.php
@@ -25,7 +25,7 @@ namespace Centreon\Infrastructure\Monitoring\ServiceGroup;
 use Centreon\Domain\Contact\Interfaces\ContactInterface;
 use Centreon\Domain\Monitoring\ServiceGroup;
 use Centreon\Infrastructure\Repository\AbstractRepositoryDRB;
-use Centreon\Domain\Security\AccessGroup;
+use Core\Security\Domain\AccessGroup\Model\AccessGroup;
 use Centreon\Domain\Entity\EntityCreator;
 use Centreon\Domain\Monitoring\ServiceGroup\Interfaces\ServiceGroupRepositoryInterface;
 use Centreon\Infrastructure\DatabaseConnection;
