+++ src/Core/Application/RealTime/UseCase/FindMetaService/FindMetaService.php
@@ -24,12 +24,12 @@ namespace Core\Application\RealTime\UseCase\FindMetaService;
 
 use Centreon\Domain\Log\LoggerTrait;
 use Core\Domain\RealTime\Model\Downtime;
-use Centreon\Domain\Security\AccessGroup;
+use Core\Security\Domain\AccessGroup\Model\AccessGroup;
 use Core\Domain\RealTime\Model\MetaService;
 use Core\Domain\RealTime\Model\Acknowledgement;
 use Core\Application\Common\UseCase\NotFoundResponse;
 use Centreon\Domain\Contact\Interfaces\ContactInterface;
-use Centreon\Domain\Security\Interfaces\AccessGroupRepositoryInterface;
+use Core\Security\Application\Repository\AccessGroupRepositoryInterface;
 use Core\Application\RealTime\Repository\ReadDowntimeRepositoryInterface;
 use Core\Application\RealTime\Repository\ReadMetaServiceRepositoryInterface;
 use Core\Domain\Configuration\Model\MetaService as MetaServiceConfiguration;
