+++ tests/php/Centreon/DependencyInjector.php
@@ -52,6 +52,7 @@ class DependencyInjector implements ArrayAccess
      * @param string|int $offset
      * @param mixed $value
      */
+    #[\ReturnTypeWillChange]
     public function offsetSet($offset, $value)
     {
         if (is_null($offset)) {
@@ -65,6 +66,7 @@ class DependencyInjector implements ArrayAccess
      * @param string|int $offset
      * @return mixed
      */
+    #[\ReturnTypeWillChange]
     public function offsetExists($offset)
     {
         return isset($this->container[$offset]);
@@ -73,6 +75,7 @@ class DependencyInjector implements ArrayAccess
     /**
      * @param string|int $offset
      */
+    #[\ReturnTypeWillChange]
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
@@ -84,6 +87,7 @@ class DependencyInjector implements ArrayAccess
      * @param string|int $offset
      * @return mixed
      */
+    #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
         return isset($this->container[$offset]) ? $this->container[$offset] : null;
