+++ tests/php/Centreon/Domain/Check/CheckServiceTest.php
@@ -28,7 +28,7 @@ use Centreon\Domain\Check\CheckService;
 use Centreon\Domain\Check\Check;
 use Centreon\Domain\Monitoring\Resource;
 use Centreon\Domain\Monitoring\Interfaces\MonitoringRepositoryInterface;
-use Centreon\Domain\Security\Interfaces\AccessGroupRepositoryInterface;
+use Core\Security\Application\Repository\AccessGroupRepositoryInterface;
 use Centreon\Domain\Engine\Interfaces\EngineServiceInterface;
 use Centreon\Domain\Entity\EntityValidator;
 use Symfony\Component\Validator\ConstraintViolationList;
