+++ www/class/centreonDB.class.php
@@ -294,6 +294,7 @@ class CentreonDB extends \PDO
      * @param mixed $parameters
      * @param mixed $parametersArgs
      */
+    #[\ReturnTypeWillChange]
     public function query($queryString, $parameters = null, ...$parametersArgs)
     {
         if (!is_null($parameters) && !is_array($parameters)) {
