+++ www/class/centreonDBStatement.class.php
@@ -74,6 +74,7 @@ class CentreonDBStatement extends \PDOStatement
      *
      * {@inheritDoc}
      */
+    #[\ReturnTypeWillChange]
     public function fetch($fetch_style = null, $cursor_orientation = PDO::FETCH_ORI_NEXT, $cursor_offset = 0)
     {
         if (is_null($this->allFetched)) {
